A Butt Ton of Code Refactoring
Need to refactor the code and change files around
