+++ BackendWork/models.py
@@ -1,0 +1,5 @@
+from django.db import models
+
+# Create your models here.
+class MyModel(models.Model):
+    pass
