+++ EcommercePlatform/settings.py
@@ -1,5 +1,5 @@
 """
-Django settings for EcommercePlatform project.
+Django settings for Crap project.
 
 Generated by 'django-admin startproject' using Django 5.0.1.
 
@@ -37,6 +37,7 @@ INSTALLED_APPS = [
     'django.contrib.sessions',
     'django.contrib.messages',
     'django.contrib.staticfiles',
+    "BackendWork.apps.BackendWorkConfig"
 ]
 
 MIDDLEWARE = [
