+++ EcommercePlatform/urls.py
@@ -1,5 +1,5 @@
 """
-URL configuration for EcommercePlatform project.
+URL configuration for Crap project.
 
 The `urlpatterns` list routes URLs to views. For more information please see:
     https://docs.djangoproject.com/en/5.0/topics/http/urls/
@@ -16,7 +16,7 @@ Including another URLconf
 """
 from django.contrib import admin
 from django.urls import path, include
-from EcommercePlatform.views import *
+from BackendWork.views import *
 
 urlpatterns = [
     path('admin/', admin.site.urls),
